precomputes Turbine trees speculatively before shreds arrive (#5014)
The commit maintains number of shreds observed in each slot within a
rolling window. If the channel is empty and there are no shreds to
retransmit, it uses the idle cycles to speculatively precompute turbine tree
for slots which have received most number of shreds within the rolling
window.

+++ turbine/src/cluster_nodes.rs
@@ -226,7 +226,7 @@ impl ClusterNodes<RetransmitStage> {
         shred: &ShredId,
         fanout: usize,
         socket_addr_space: &SocketAddrSpace,
-    ) -> Result<(/*root_distance:*/ usize, Vec<SocketAddr>), Error> {
+    ) -> Result<(/*root_distance:*/ u8, Vec<SocketAddr>), Error> {
         // Exclude slot leader from list of nodes.
         if slot_leader == &self.pubkey {
             return Err(Error::Loopback {
@@ -596,7 +596,7 @@ pub(crate) fn get_broadcast_protocol(_: &ShredId) -> Protocol {
 }
 
 #[inline]
-fn get_root_distance(index: usize, fanout: usize) -> usize {
+fn get_root_distance(index: usize, fanout: usize) -> u8 {
     if index == 0 {
         0
     } else if index <= fanout {
